Refactor, get rid of 'Configurations' wip-component
