--- src/renderer/components/Configurations.css
@@ -1,4 +1,0 @@
-.ConfigurationsWrapper {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-}
