+++ src/renderer/index.html
@@ -15,8 +15,6 @@
     <link rel="stylesheet" type="text/css" href="components/App.css">
     <script type="text/babel" src="components/Settings.jsx"></script>
     <link rel="stylesheet" type="text/css" href="components/Settings.css">
-    <script type="text/babel" src="components/Configurations.jsx"></script>
-    <link rel="stylesheet" type="text/css" href="components/Configurations.css">
     <script type="text/babel" src="components/RunConfiguration.jsx"></script>
     <link rel="stylesheet" type="text/css" href="components/RunConfiguration.css">
     <script type="text/babel" src="components/ScenarioConfiguration.jsx"></script>
